requirements.txt
pandas==2.1.4
numpy==1.23.5
wandb==0.16.1
tqdm==4.66.1
pytorch_lightning==2.1.2
transformers[torch]>=4.41.0
rouge==1.0.1
jupyter==1.0.0
jupyterlab==4.0.9
python-dotenv==1.0.1
gradio==5.38.2
seaborn
nltk
konlpy
scikit-learn
umap-learn
hdbscan
sentencepiece==0.2.0
peft==0.16.0
datasets==4.0.0
accelerate>=0.24.0
bitsandbytes>=0.46.1